fix main

+++ src/main.rs
@@ -35,7 +35,7 @@ fn write_to_csv(filepath: &str, header: &str, data: Vec<(i32, i32, i32, i32, i32
     Ok(())
 }
 
-fn simulate_hoyo_games(num_simulations: usize) -> io::Result<()> {
+fn simulate_hoyo_games(num_simulations: u64) -> io::Result<()> {
     let num_threads = num_cpus::get();
     let num_simulations_per_thread = num_simulations / num_threads;
